more UI work on the main menu
I fixed up the formatting of the main menu and now understand how Bevy UI works a lot better than before. for the next work on the main menu, I plan on adding more buttons with different characters. That will come later, however. Next item on the agenda is making a UI for in-game, which will let you see the character's health and experience (though exp is still not implemented yet). Also, I need to work on the pause menu which should be tied to a key on the keyboard.